Start splitting ring-backed implementation into own crate
