+++ oxide-auth-actix/examples/actix-example/Cargo.toml
@@ -12,6 +12,7 @@ actix-web = "1.0"
 env_logger = "0.6"
 futures = "0.1"
 oxide-auth = { version = "0.5.0-preview.0", path = "./../../../oxide-auth" }
+oxide-auth-ring = { version = "0.5.0-preview.0", path = "./../../../oxide-auth-ring" }
 oxide-auth-actix = { version = "0.1.0", path = "./../../" }
 reqwest = "=0.9.5"
 serde = "1.0"
