+++ oxide-auth-ring/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "oxide-auth"
+name = "oxide-auth-ring"
 version = "0.5.0-preview.0"
 authors = ["Andreas Molzer <andreas.molzer@gmx.de>"]
 repository = "https://github.com/HeroicKatora/oxide-auth.git"
@@ -11,22 +11,17 @@ keywords = ["oauth", "server", "oauth2"]
 categories = ["web-programming::http-server", "authentication"]
 license = "MIT"
 
-autoexamples = false
+edition = "2018"
 
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 base64 = "0.9"
 chrono = "0.4.2"
+oxide-auth = { version = "0.5.0-preview.0", path = "../oxide-auth" }
+ring = ">=0.13,<0.15"
+rmp-serde = "^0.13"
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
-rmp-serde = "^0.13"
 url = "1.7"
-
-[features]
-
-[dev-dependencies]
-reqwest = "^0.9"
-
-[package.metadata.docs.rs]
-features = []
