+++ oxide-auth-ring/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod authorizer;
+pub mod generator;
+pub mod issuer;
+pub mod pkce;
+pub mod registrar;
+#[cfg(test)]
+mod tests;
