+++ oxide-auth-ring/src/tests/refresh.rs
@@ -1,7 +1,11 @@
-use primitives::issuer::{Issuer, IssuedToken, RefreshedToken, TokenMap};
-use primitives::generator::RandomGenerator;
-use primitives::grant::{Grant, Extensions};
-use primitives::registrar::{Client, ClientMap};
+use oxide_auth::{
+    primitives::{
+        issuer::{Issuer, IssuedToken, RefreshedToken, TokenMap},
+        grant::{Grant, Extensions},
+        registrar::{Client, ClientMap},
+    },
+    frontends::simple::endpoint::{refresh_flow, resource_flow},
+};
 
 use std::collections::HashMap;
 
@@ -11,7 +15,8 @@ use serde_json;
 
 use super::{Body, CraftedRequest, CraftedResponse, Status, ToSingleValueQuery};
 use super::defaults::*;
-use frontends::simple::endpoint::{refresh_flow, resource_flow};
+
+use crate::{registrar::Pbkdf2, generator::RandomGenerator};
 
 struct RefreshTokenSetup {
     registrar: ClientMap,
@@ -23,6 +28,7 @@ struct RefreshTokenSetup {
 impl RefreshTokenSetup {
     fn private_client() -> Self {
         let mut registrar = ClientMap::new();
+        registrar.set_password_policy(Pbkdf2::default());
         let mut issuer = TokenMap::new(RandomGenerator::new(16));
 
         let client = Client::confidential(EXAMPLE_CLIENT_ID,
@@ -39,7 +45,7 @@ impl RefreshTokenSetup {
             extensions: Extensions::new(),
         };
 
-        registrar.register_client(client);
+        registrar.register_client(client).unwrap();
         let issued = issuer.issue(grant).unwrap();
         assert!(!issued.refresh.is_empty());
 
@@ -57,6 +63,7 @@ impl RefreshTokenSetup {
 
     fn public_client() -> Self {
         let mut registrar = ClientMap::new();
+        registrar.set_password_policy(Pbkdf2::default());
         let mut issuer = TokenMap::new(RandomGenerator::new(16));
 
         let client = Client::public(EXAMPLE_CLIENT_ID,
@@ -72,7 +79,7 @@ impl RefreshTokenSetup {
             extensions: Extensions::new(),
         };
 
-        registrar.register_client(client);
+        registrar.register_client(client).unwrap();
         let issued = issuer.issue(grant).unwrap();
         assert!(!issued.refresh.is_empty());
 
@@ -233,7 +240,7 @@ fn public_private_invalid_grant() {
             EXAMPLE_REDIRECT_URI.parse().unwrap(),
             EXAMPLE_SCOPE.parse().unwrap(),
             EXAMPLE_PASSPHRASE.as_bytes());
-    setup.registrar.register_client(client);
+    setup.registrar.register_client(client).unwrap();
 
     let basic_authorization = base64::encode(&format!("{}:{}",
         "PrivateClient", EXAMPLE_PASSPHRASE));
