+++ oxide-auth/src/lib.rs
@@ -73,7 +73,6 @@
 extern crate base64;
 extern crate chrono;
 extern crate url;
-extern crate ring;
 extern crate rmp_serde;
 extern crate serde;
 #[macro_use]
