+++ oxide-auth/src/primitives/mod.rs
@@ -34,7 +34,7 @@
 
 use chrono::DateTime;
 use chrono::Utc;
-use url::Url;
+pub use url::Url;
 
 pub mod authorizer;
 pub mod generator;
@@ -43,13 +43,14 @@ pub mod issuer;
 pub mod registrar;
 pub mod scope;
 
-type Time = DateTime<Utc>;
+/// Alias for DateTime<Utc>
+pub type Time = DateTime<Utc>;
 
 /// Commonly used primitives for frontends and backends.
 pub mod prelude {
     pub use super::authorizer::{Authorizer, AuthMap};
-    pub use super::issuer::{IssuedToken, Issuer, TokenMap, TokenSigner};
-    pub use super::generator::{Assertion, TagGrant, RandomGenerator};
+    pub use super::issuer::{IssuedToken, Issuer, TokenMap};
+    pub use super::generator::TagGrant;
     pub use super::registrar::{Registrar, Client, ClientUrl, ClientMap, PreGrant};
     pub use super::scope::Scope;
 }
